feat: Updated README.md for accuracy and clarity; restructured Docker containers to improve scalability and prepared the groundwork for Kubernetes integration
