--- scraping-service/libft/__init__.py
@@ -1,50 +1,0 @@
-# Import from libft_utility.py 
-from .libft_utility import (
-	wait_for_element_conad,
-	wait_for_element,
-	wait_for_elements_conad,
-	wait_for_elements,
-	fetch_data,
-	extract_micro_categories,
-	update_env_with_dotenv,
-	to_float,
-	get_html_from_url,
-)
-
-# Import from send_data.py
-from .send_data import (
-	send_data_to_receiver,
-	create_store,
-	get_all_stores,
-	get_store_by_grocery_and_city,
-)
-
-# Import from libft_gros.py
-from .libft_gros import (
-	get_html_from_url,
-	has_questo_indirizzo,
-	has_superficie,
-	has_phone_number,
-	geocode,
-)
-
-# Define what is exposed when `from libft import *` is used
-__all__ = [
-	"wait_for_element_conad",
-	"wait_for_element",
-	"wait_for_elements_conad",
-	"wait_for_elements",
-	"fetch_data",
-	"extract_micro_categories",
-	"update_env_with_dotenv",
-	"to_float",
-	"get_html_from_url",
-	"has_questo_indirizzo",
-	"has_superficie",
-	"has_phone_number",
-	"geocode",
-	"send_data_to_receiver",
-	"create_store",
-	"get_all_stores",
-	"get_store_by_grocery_and_city",
-]
