--- scraping-service/oasi_tigre_shop.json
@@ -1,56 +1,0 @@
-[
-    {
-        "name": "Supermercato Tigre",
-        "lat": 41.8412735,
-        "long": 12.8720899,
-        "street": "Via Pedemontana, 88, 00036 PALESTRINA RM",
-        "city": "PALESTRINA",
-        "working_hours": "{'mon': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'tue': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'wed': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'thu': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'fri': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'sat': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'sun': {'morningOpen': '09:00', 'morningClose': '13:00', 'eveningOpen': '', 'eveningClose': ''}}",
-        "picks_up_in_shop": true
-    },
-    {
-        "name": "Supermercato Tigre",
-        "lat": 41.959978,
-        "long": 12.5351033,
-        "street": "Via T. Boetti Valvassura, 110, 00139 Roma RM",
-        "city": "Roma",
-        "working_hours": "{'mon': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'tue': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'wed': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'thu': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'fri': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'sat': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'sun': {'morningOpen': '08:00', 'morningClose': '13:30', 'eveningOpen': '', 'eveningClose': ''}}",
-        "picks_up_in_shop": true
-    },
-    {
-        "name": "Supermercato Tigre",
-        "lat": 41.93363830000001,
-        "long": 12.5130374,
-        "street": "Via Montebuono, 8, 00199 Roma RM",
-        "city": "Roma",
-        "working_hours": "{'mon': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'tue': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'wed': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'thu': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'fri': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'sat': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'sun': {'morningOpen': '08:30', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}}",
-        "picks_up_in_shop": true
-    },
-    {
-        "name": "Supermercato Tigre",
-        "lat": 41.9268843,
-        "long": 12.5156112,
-        "street": "Viale Eritrea, 75, 00199 Roma RM",
-        "city": "Roma",
-        "working_hours": "{'mon': {'morningOpen': '08:00', 'morningClose': '20:30', 'eveningOpen': '', 'eveningClose': ''}, 'tue': {'morningOpen': '08:00', 'morningClose': '20:30', 'eveningOpen': '', 'eveningClose': ''}, 'wed': {'morningOpen': '08:00', 'morningClose': '20:30', 'eveningOpen': '', 'eveningClose': ''}, 'thu': {'morningOpen': '08:00', 'morningClose': '20:30', 'eveningOpen': '', 'eveningClose': ''}, 'fri': {'morningOpen': '08:00', 'morningClose': '20:30', 'eveningOpen': '', 'eveningClose': ''}, 'sat': {'morningOpen': '08:00', 'morningClose': '20:30', 'eveningOpen': '', 'eveningClose': ''}, 'sun': {'morningOpen': '09:00', 'morningClose': '13:30', 'eveningOpen': '', 'eveningClose': ''}}",
-        "picks_up_in_shop": true
-    },
-    {
-        "name": "Supermercato Tigre",
-        "lat": 41.9215852,
-        "long": 12.4943683,
-        "street": "Viale Liegi, 29, 00198 Roma RM",
-        "city": "Roma",
-        "working_hours": "{'mon': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'tue': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'wed': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'thu': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'fri': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'sat': {'morningOpen': '08:00', 'morningClose': '21:00', 'eveningOpen': '', 'eveningClose': ''}, 'sun': {'morningOpen': '08:30', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}}",
-        "picks_up_in_shop": true
-    },
-    {
-        "name": "Supermercato Tigre",
-        "lat": 41.80154139999999,
-        "long": 12.5980664,
-        "street": "Piazza Trento e Trieste, 11, 00043 Ciampino RM",
-        "city": "Ciampino",
-        "working_hours": "{'mon': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'tue': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'wed': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'thu': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'fri': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'sat': {'morningOpen': '08:00', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}, 'sun': {'morningOpen': '08:30', 'morningClose': '20:00', 'eveningOpen': '', 'eveningClose': ''}}",
-        "picks_up_in_shop": true
-    }
-]
