+++ scraping-service/scraping_real/Dockerfile.scraping_gros_groups
@@ -1,0 +1,17 @@
+# Base Python image
+FROM python:3.9-slim
+
+# Set working directory
+WORKDIR /app/gros_groups
+
+# Copy the gros_groups folder
+COPY scraping/gros_groups /app/gros_groups
+COPY libft /app/libft
+
+# Install Python dependencies
+COPY requirements.txt /app/
+RUN pip install --no-cache-dir -r /app/requirements.txt
+
+# Run the script
+CMD ["python3", "scheduler.py"]
+
