--- scraping-service/scraping_real/Dockerfile.shop
@@ -1,26 +1,0 @@
-FROM python:3.10
-
-WORKDIR /app
-
-COPY requirements.txt .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY libft ./libft
-
-# COPY scraping_cts.py ./scraping_cts.py
-
-COPY scraping_cts_shop.py ./scraping_cts_shop.py
-
-COPY .env .env
-
-ENV PYTHONPATH=/app
-
-# RUN chmod +x scraping_cts.py
-
-RUN chmod +x scraping_cts_shop.py
-
-CMD ["python3", "scraping_cts_shop.py"]
-
-# CMD ["python3", "scraping_cts.py"]
-
