+++ scraping-service/scraping_real/requirements.txt
@@ -1,4 +1,6 @@
-beautifulsoup4
 requests
 selenium
 python-dotenv
+beautifulsoup4
+undetected_chromedriver
+
