+++ scraping-service/scraping_real/scraping/conad/scraping_conad.py
@@ -12,9 +12,7 @@ from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.support import expected_conditions as EC
 
 sys.path.append('../')
-from libft import wait_for_element, wait_for_elements, wait_for_element_conad, wait_for_elements_conad, update_env_with_dotenv, to_float, get_store_by_grocery_and_city, send_data_to_receiver
-
-load_dotenv()
+from libft import wait_for_element, wait_for_elements, wait_for_element_conad, wait_for_elements_conad, to_float, get_store_by_grocery_and_city, send_data_to_receiver
 
 # Finds product information on a webpage and sends it to a specified function.
 # Returns: The number of products processed successfully.
@@ -120,90 +118,83 @@ def change_shop_location(driver, location, wait):
 
 if __name__ == "__main__":
 
-	all_shop = os.environ.get("ALL_SHOP")
-	count_shop = os.environ.get("COUNT_SHOP")
+	for count_shop in range(0, 61):
 
-	result1 = get_store_by_grocery_and_city("Conad Superstore", "ROMA")
-	result2 = get_store_by_grocery_and_city("Conad", "ROMA")
-	result3 = get_store_by_grocery_and_city("Conad City", "ROMA")
-	result4 = get_store_by_grocery_and_city("Spazio Conad", "ROMA")
-	result5 = get_store_by_grocery_and_city("Conad", "GUIDONIA MONTECELIO")
+		result1 = get_store_by_grocery_and_city("Conad Superstore", "ROMA")
+		result2 = get_store_by_grocery_and_city("Conad", "ROMA")
+		result3 = get_store_by_grocery_and_city("Conad City", "ROMA")
+		result4 = get_store_by_grocery_and_city("Spazio Conad", "ROMA")
+		result5 = get_store_by_grocery_and_city("Conad", "GUIDONIA MONTECELIO")
 
-	shop_list = {'stores': result1['stores'] + result2['stores'] + result3['stores'] + result4['stores'] + result5['stores']}
+		shop_list = {'stores': result1['stores'] + result2['stores'] + result3['stores'] + result4['stores'] + result5['stores']}
 
-	if int(count_shop) == 61:
-		update_env_with_dotenv(".env", "COUNT_SHOP", 0)
-		count_shop = 0
-	
-	count_shop = int(count_shop)
+		shop = shop_list['stores'][count_shop]
 
-	shop = shop_list['stores'][count_shop]
+		options = uc.ChromeOptions()
 
-	options = uc.ChromeOptions()
+		options.binary_location = "/usr/bin/google-chrome"
 
-	options.binary_location = "/usr/bin/google-chrome"
+		options.add_argument("--headless")
+		options.add_argument("--no-sandbox")
+		options.add_argument("--disable-dev-shm-usage")
 
-	options.add_argument("--headless")
-	options.add_argument("--no-sandbox")
-	options.add_argument("--disable-dev-shm-usage")
+		# Initialize the undetected Chrome driver
+		driver = uc.Chrome(options=options, use_subprocess=False)
+		wait = WebDriverWait(driver, 30)
 
-	# Initialize the undetected Chrome driver
-	driver = uc.Chrome(options=options, use_subprocess=False)
-	wait = WebDriverWait(driver, 30)
+		total_items_processed = 0
 
-	total_items_processed = 0
-	product_list = []
-	change_shop_location(driver, shop['street'], wait)
+		change_shop_location(driver, shop['street'], wait)
 
-	wait.until(EC.visibility_of_element_located((By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a')))
-	driver.find_element(By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a').click()
-	time.sleep(3)
-
-	count_macro_cat = len((wait_for_elements(driver, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li')))
-
-	time.sleep(3)
-
-	for macro_count in range(1, count_macro_cat + 1):
-		wait.until(EC.visibility_of_element_located((By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li[{macro_count}]/a')))
-		driver.find_element(By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li[{macro_count}]/a').click()
-
-		count_micro_cat = len((wait_for_elements(driver, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[2]/div[{macro_count}]/ul/li')))
+		wait.until(EC.visibility_of_element_located((By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a')))
+		driver.find_element(By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a').click()
 		time.sleep(3)
-	
-		for micro_count in range(2, count_micro_cat + 1):
-			wait.until(EC.visibility_of_element_located((By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[2]/div[{macro_count}]/ul/li[{micro_count}]/a')))
-			driver.find_element(By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[2]/div[{macro_count}]/ul/li[{micro_count}]/a').click()
-			time.sleep(3)
 
-			if (wait_for_element_conad(driver, "/html/body/main/div/div[2]/div[2]/div[4]/section") == None):
-				advertising = 4
-			else:
-				advertising = 5
+		count_macro_cat = len((wait_for_elements(driver, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li')))
 
-			count_pages = len((wait_for_elements(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li')))
-		
-			if (count_pages > 1):
-				n_pages = (wait_for_element(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li[{count_pages - 1}]')).text
-			else:
-				n_pages = 1
-		
-			while True:
-				counter = int(len((wait_for_elements(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li'))))
-		
-				if (wait_for_elements_conad(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li[{counter}]/a/span') == None):
-					break
-		
-				n_cards = len(wait_for_elements(driver, f'//*[@id="#top"]/div/div[2]/div[2]/div[{advertising}]/div'))
+		time.sleep(3)
 
-				total_items_processed = find_and_send_info(driver, n_cards, advertising, shop)
-
-				driver.find_element(By.XPATH, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li[{counter}]/a').click()
-
-			wait.until(EC.visibility_of_element_located((By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a')))
-			driver.find_element(By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a').click()
-			time.sleep(3)
+		for macro_count in range(1, count_macro_cat + 1):
 			wait.until(EC.visibility_of_element_located((By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li[{macro_count}]/a')))
 			driver.find_element(By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li[{macro_count}]/a').click()
-		
-	print(total_items_processed)
-	driver.quit()
+
+			count_micro_cat = len((wait_for_elements(driver, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[2]/div[{macro_count}]/ul/li')))
+			time.sleep(3)
+
+			for micro_count in range(2, count_micro_cat + 1):
+				wait.until(EC.visibility_of_element_located((By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[2]/div[{macro_count}]/ul/li[{micro_count}]/a')))
+				driver.find_element(By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[2]/div[{macro_count}]/ul/li[{micro_count}]/a').click()
+				time.sleep(3)
+
+				if (wait_for_element_conad(driver, "/html/body/main/div/div[2]/div[2]/div[4]/section") == None):
+					advertising = 4
+				else:
+					advertising = 5
+
+				count_pages = len((wait_for_elements(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li')))
+			
+				if (count_pages > 1):
+					n_pages = (wait_for_element(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li[{count_pages - 1}]')).text
+				else:
+					n_pages = 1
+			
+				while True:
+					counter = int(len((wait_for_elements(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li'))))
+			
+					if (wait_for_elements_conad(driver, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li[{counter}]/a/span') == None):
+						break
+			
+					n_cards = len(wait_for_elements(driver, f'//*[@id="#top"]/div/div[2]/div[2]/div[{advertising}]/div'))
+
+					total_items_processed = find_and_send_info(driver, n_cards, advertising, shop)
+
+					driver.find_element(By.XPATH, f'/html/body/main/div/div[2]/div[2]/div[{advertising + 1}]/ul/li[{counter}]/a').click()
+
+				wait.until(EC.visibility_of_element_located((By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a')))
+				driver.find_element(By.XPATH, '/html/body/section[1]/header/div/nav[2]/ul/li[1]/a').click()
+				time.sleep(3)
+				wait.until(EC.visibility_of_element_located((By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li[{macro_count}]/a')))
+				driver.find_element(By.XPATH, f'/html/body/section[1]/header/div/nav[2]/ul/li[1]/div/div/div/div[1]/ul/li[{macro_count}]/a').click()
+			
+		print(total_items_processed)
+		driver.quit()
