--- scraping-service/scraping_real/scraping/requirements.txt
@@ -1,5 +1,0 @@
-geocoder
-requests
-selenium
-python-dotenv
-undetected-chromedriver
